+++ src/main/java/net/silentchaos512/gems/item/PetSummonerItem.java
@@ -4,7 +4,6 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.network.chat.Component;
 import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.EntityType;
@@ -12,28 +11,19 @@ import net.minecraft.world.entity.TamableAnimal;
 import net.minecraft.world.entity.animal.Cat;
 import net.minecraft.world.entity.animal.Wolf;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.silentchaos512.gems.SilentGems;
-import net.silentchaos512.gems.util.TextUtil;
 
-import java.util.List;
 import java.util.function.Function;
 
-public class PetSummonerItem extends Item {
+public class PetSummonerItem extends ItemWithFlavorText {
     private final Function<Level, ? extends TamableAnimal> petFactory;
 
     public PetSummonerItem(Function<Level, ? extends TamableAnimal> petFactory, Properties properties) {
         super(properties);
         this.petFactory = petFactory;
-    }
-
-    @Override
-    public void appendHoverText(ItemStack stack, TooltipContext tooltipContext, List<Component> tooltip, TooltipFlag flagIn) {
-        tooltip.add(TextUtil.itemSub(this, "desc"));
     }
 
     @Override