+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21.1-5.1.1] - 2025-06-22
+### Added
+- Sparkling bone meal, which can be used to grow random glowroses (related to #588)
+### Fixed
+- Bone meal creating only ruby glowroses when used [#588]
+
 ## [1.21.1-5.1.0] - 2025-06-16
 ### Added
 - Teleporters, redstone teleporters, and teleporter anchors are back!